raphael: Merge FOD HAL stuff to fingerprint HAL
 * Also nuke deprecated Lineage FOD HAL.
