+++ device.mk
@@ -206,8 +206,7 @@ PRODUCT_PACKAGES += \
 
 # Fingerprint
 PRODUCT_PACKAGES += \
-    android.hardware.biometrics.fingerprint@2.3-service.xiaomi_raphael \
-    vendor.lineage.biometrics.fingerprint.inscreen@1.0-service.xiaomi_raphael
+    android.hardware.biometrics.fingerprint@2.3-service.xiaomi_raphael
 
 # FM
 PRODUCT_PACKAGES += \
