+++ fingerprint/Android.bp
@@ -29,7 +29,19 @@ cc_binary {
         "android.hardware.biometrics.fingerprint@2.1",
         "android.hardware.biometrics.fingerprint@2.2",
         "android.hardware.biometrics.fingerprint@2.3",
+        "vendor.goodix.hardware.biometrics.fingerprint@2.1",
         "vendor.xiaomi.hardware.fingerprintextension@1.0",
     ],
     proprietary: true,
+}
+
+cc_library_static {
+    name: "libudfps_extension.xiaomi_raphael",
+    srcs: ["UdfpsExtension.cpp"],
+    include_dirs: [
+        "frameworks/native/services/surfaceflinger/CompositionEngine/include"
+    ],
+    header_libs: [
+        "generated_kernel_headers",
+    ],
 }
