+++ fingerprint/UdfpsExtension.cpp
@@ -16,9 +16,9 @@
 
 #include <stdint.h>
 #include <drm/sde_drm.h>
-#include <compositionengine/FodExtension.h>
+#include <compositionengine/UdfpsExtension.h>
 
-uint32_t getFodZOrder(uint32_t z, bool touched) {
+uint32_t getUdfpsZOrder(uint32_t z, bool touched) {
     if (touched) {
         z |= FOD_PRESSED_LAYER_ZORDER;
     }
@@ -26,7 +26,7 @@ uint32_t getFodZOrder(uint32_t z, bool touched) {
     return z;
 }
 
-uint64_t getFodUsageBits(uint64_t usageBits, bool touched) {
+uint64_t getUdfpsUsageBits(uint64_t usageBits, bool touched) {
     (void) touched;
     return usageBits;
 }
