+++ fingerprint/android.hardware.biometrics.fingerprint@2.3-service.xiaomi_raphael.rc
@@ -10,6 +10,11 @@ on init
     chown system system /dev/fortsense_fp
 
 on boot
+    chown system system /sys/devices/platform/soc/soc:qcom,dsi-display-primary/fod_ui
+    chown system system /sys/devices/virtual/touch/tp_dev/fod_status
+    chmod 0660 /sys/devices/platform/soc/soc:qcom,dsi-display-primary/fod_ui
+    chmod 0660 /sys/devices/virtual/touch/tp_dev/fod_status
+
     chown system system /sys/bus/platform/devices/soc:fingerprint_fpc/irq
     chown system system /sys/bus/platform/devices/soc:fingerprint_fpc/irq_enable
     chown system system /sys/bus/platform/devices/soc:fingerprint_fpc/wakeup_enable
