--- fod/vendor.lineage.biometrics.fingerprint.inscreen@1.0-service.xiaomi_raphael.rc
@@ -1,13 +1,0 @@
-on boot
-    chown system system /sys/devices/platform/soc/soc:qcom,dsi-display-primary/fod_ui
-    chown system system /sys/devices/virtual/touch/tp_dev/fod_status
-
-    chmod 0660 /sys/devices/platform/soc/soc:qcom,dsi-display-primary/fod_ui
-    chmod 0660 /sys/devices/virtual/touch/tp_dev/fod_status
-
-service fingerprint-inscreen-1-0 /vendor/bin/hw/vendor.lineage.biometrics.fingerprint.inscreen@1.0-service.xiaomi_raphael
-    interface vendor.lineage.biometrics.fingerprint.inscreen@1.0::IFingerprintInscreen default
-    class hal
-    user system
-    group system
-    shutdown critical
